Reject he-aac on Android; fail over to lc-aac via source option
Add re-usable dashCanPlay function; do not make spurious check for
engine source option (which is done in core).

It's unclear whether or when this fix
https://chromium.googlesource.com/chromium/src.git/+/0b5ec458acf03e3507a3737cfc483df0694cf803%5E!/
is available in actual devices, and what effect it has.

Therefore bail out on Android if mp4a.40.5 is set in codecs.
But offer a failover by extending the dash.type and dash.codecs options
as dashType and dashCodecs source options.
These options are of advanced category, as setting them without a
fallback Dash stream prevents the player from loading.
See also:
https://github.com/flowplayer/flowplayer-mpegdash/issues/13